+++ app/src/main/java/com/xacalet/moobies/db/model/UserRatingDbModel.kt
@@ -6,5 +6,5 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "user_rating")
 data class UserRatingDbModel(
     @PrimaryKey val id: Long,
-    val stars: Byte
+    val stars: Int
 )
