+++ app/src/main/java/com/xacalet/moobies/presentation/userrating/UserRatingUiModel.kt
@@ -3,6 +3,6 @@ package com.xacalet.moobies.presentation.userrating
 data class UserRatingUiModel(
     val showId: Long,
     val showName: String,
-    val stars: Byte?,
+    val stars: Int?,
     val poserImageUrl: String?
 )
