+++ domain/src/main/java/com/xacalet/domain/repository/UserRatingRepository.kt
@@ -4,13 +4,13 @@ import kotlinx.coroutines.flow.Flow
 
 interface UserRatingRepository {
 
-    suspend fun addUserRating(id: Long, stars: Byte)
+    suspend fun addUserRating(id: Long, stars: Int)
 
     suspend fun deleteUserRating(id: Long)
 
-    fun getUserRatingFlow(id: Long): Flow<Byte?>
+    fun getUserRatingFlow(id: Long): Flow<Int?>
 
-    suspend fun getUserRating(id: Long): Byte?
+    suspend fun getUserRating(id: Long): Int?
 
-    suspend fun getTitlesByRating(rating: Byte): List<Long>
+    suspend fun getTitlesByRating(rating: Int): List<Long>
 }
