+++ domain/src/main/java/com/xacalet/domain/usecase/GetUserRatingFlowUseCase.kt
@@ -8,5 +8,5 @@ class GetUserRatingFlowUseCase @Inject constructor(
     private val userRatingRepository: UserRatingRepository
 ) {
 
-    operator fun invoke(id: Long): Flow<Byte?> = userRatingRepository.getUserRatingFlow(id)
+    operator fun invoke(id: Long): Flow<Int?> = userRatingRepository.getUserRatingFlow(id)
 }
